separate ray stuff into its own file
