+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -12,7 +12,6 @@ import json
 import logging
 import os
 import re
-import sys
 from collections import Counter, defaultdict
 import typing
 import uuid
@@ -33,7 +32,6 @@ from typing import (
     Tuple,
 )
 
-from modin.config import context as config_context
 import modin.pandas as pd
 from modin.pandas import Series, DataFrame
 from modin.pandas.base import BasePandasDataset
@@ -45,7 +43,7 @@ import pandas.io.parsers
 from pandas.core.interchange.dataframe_protocol import DataFrame as InterchangeDataframe
 import pandas.io.parsers.readers
 import pytz  # type: ignore
-from modin.core.storage_formats import BaseQueryCompiler, PandasQueryCompiler  # type: ignore
+from modin.core.storage_formats import BaseQueryCompiler  # type: ignore
 from modin.core.storage_formats.base.query_compiler import QCCoercionCost
 from pandas import Timedelta
 from pandas._libs import lib
@@ -100,7 +98,6 @@ from snowflake.snowpark._internal.utils import (
     generate_random_alphanumeric,
     parse_table_name,
     random_name_for_temp_object,
-    is_interactive,
 )
 from snowflake.snowpark.column import CaseExpr, Column as SnowparkColumn
 from snowflake.snowpark.dataframe import DataFrame as SnowparkDataFrame
@@ -408,6 +405,10 @@ from snowflake.snowpark.modin.utils import MODIN_UNNAMED_SERIES_LABEL
 from snowflake.snowpark.modin.plugin.utils.numpy_to_pandas import (
     NUMPY_UNIVERSAL_FUNCTION_TO_SNOWFLAKE_FUNCTION,
 )
+from snowflake.snowpark.modin.plugin.compiler.ray_utils import (
+    move_from_ray_helper,
+    move_to_ray_helper,
+)
 from snowflake.snowpark.row import Row
 from snowflake.snowpark.session import Session
 from snowflake.snowpark.types import (
@@ -564,41 +565,6 @@ def _propagate_attrs_on_methods(cls):  # type: ignore
         elif inspect.isfunction(attr_value):
             setattr(cls, attr_name, propagate_attrs_decorator(attr_value))
     return cls
-
-
-class SnowflakeWriterActorBase:
-    def __init__(
-        self, table_name: str, connection_creds: dict[str, Any]
-    ) -> None:  # pragma: no cover
-        self.table_name = table_name
-        try:
-            self.session = Session.builder.configs(connection_creds).getOrCreate()
-        except Exception as e:
-            logging.error(
-                "Could not get or create Snowpark session. Ensure you have a "
-                "~/.snowflake/connections.toml file with the correct credentials. "
-                f"{e}"
-            )
-            raise RuntimeError("Could not get or create Snowpark session.") from e
-
-    def write(self, batch: native_pd.DataFrame) -> None:  # pragma: no cover
-        self.session.write_pandas(
-            df=batch,
-            table_name=self.table_name,
-            table_type="",
-            parallel=4,
-            auto_create_table=True,
-            overwrite=False,
-        )
-
-
-try:  # pragma: no cover
-    import ray  # type: ignore[import]
-    from ray.util.actor_pool import ActorPool  # type: ignore[import]
-
-    SnowflakeWriterActor = ray.remote(SnowflakeWriterActorBase)
-except ImportError:
-    SnowflakeWriterActor = None
 
 
 @_propagate_attrs_on_methods
@@ -1248,211 +1214,15 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
     def execute(self) -> None:
         pass
 
-    def _move_to_ray(self) -> PandasQueryCompiler:  # pragma: no cover
-        """
-        Move the QueryCompiler to a Ray backend using the `ml.data.DataConnector`.
-
-        Returns:
-            A new Ray-backed PandasQueryCompiler
-        """
-        try:
-            from snowflake.ml.data.data_connector import DataConnector  # type: ignore[import]
-        except ImportError as e:  # pragma: no cover
-            raise ImportError(
-                "DataConnector is required for efficient transfer to Ray. "
-                "Please `pip install snowflake-ml`."
-            ) from e
-
-        cached_qc = self.cache_result()
-        original_index_labels = self.get_index_names()
-        original_column_labels = self.columns.tolist()
-        snowflake_index_labels = [
-            label if is_interactive() else unquote_name_if_quoted(label)
-            for label in cached_qc._modin_frame.index_column_snowflake_quoted_identifiers
-        ]
-        snowflake_column_labels = [
-            label if is_interactive() else unquote_name_if_quoted(label)
-            for label in cached_qc._modin_frame.data_column_snowflake_quoted_identifiers
-        ]
-        snowpark_df = (
-            cached_qc._modin_frame.ordered_dataframe.to_projected_snowpark_dataframe(
-                sort=True
-            )
-        )
-        # Drop all metadata (ordering, row position, row count) columns
-        snowpark_df = snowpark_df.drop(
-            snowpark_df.columns[
-                len(snowflake_index_labels) + len(snowflake_column_labels) :
-            ],
-        )
-
-        data_connector = DataConnector.from_dataframe(snowpark_df)
-        ray_ds = data_connector.to_ray_dataset()
-        with config_context(Backend="Ray"):
-            ray_df = pd.io.from_ray(ray_ds)
-
-        ray_df.set_index(
-            snowflake_index_labels,
-            drop=True,
-            inplace=True,
-        )
-        ray_df.index.set_names(original_index_labels, inplace=True)
-        ray_df.columns = original_column_labels
-        return ray_df._query_compiler
-
     def move_to(
         self, target_backend: str
     ) -> Union[BaseQueryCompiler, Any]:  # pragma: no cover
         try:
             if target_backend == "Ray":
-                return self._move_to_ray()
+                return move_to_ray_helper(self)
             return NotImplemented
         except Exception:
             return NotImplemented
-
-    @classmethod
-    def _move_from_ray(
-        cls,
-        ray_qc: BaseQueryCompiler,
-        *,
-        max_sessions: int,
-    ) -> "SnowflakeQueryCompiler":  # pragma: no cover
-        """
-        Move the data from Ray to Snowflake by writing to a Snowflake table. Preserves
-        the row position order from the original dataframe.
-
-        Args:
-            ray_qc: The Ray-backed query compiler.
-            max_sessions: The maximum number of sessions to use for writes.
-
-        Returns:
-            A new SnowflakeQueryCompiler with the data
-        """
-        if SnowflakeWriterActor is None:  # pragma: no cover
-            raise ImportError(
-                "Ray is required for this operation. Please `pip install modin[ray]`."
-            )
-
-        def get_connection_creds() -> dict[str, Any]:  # pragma: no cover
-            """
-            Get the connection credentials from the notebook environment or an empty dict
-            otherwise to use default connection parameters.
-            """
-            if "snowbook" in sys.modules:
-                try:
-                    with open("/snowflake/session/token") as token_file:
-                        return {
-                            "host": os.getenv("SNOWFLAKE_HOST"),
-                            "account": os.getenv("SNOWFLAKE_ACCOUNT"),
-                            "token": token_file.read(),
-                            "authenticator": "oauth",
-                            "protocol": "https",
-                            "database": os.getenv("SNOWFLAKE_DATABASE"),
-                            "schema": os.getenv("SNOWFLAKE_SCHEMA"),
-                            "port": os.getenv("SNOWFLAKE_PORT"),
-                            "warehouse": os.getenv("SNOWFLAKE_WAREHOUSE", ""),
-                        }
-                except Exception:
-                    logging.error(
-                        "Could not read session token from notebook environment. "
-                        "Attempting to use default connection parameters."
-                    )
-                    return {}
-            return {}
-
-        connection_creds = get_connection_creds()
-        table_name = random_name_for_temp_object(TempObjectType.TABLE)
-        pool = ActorPool(
-            [
-                SnowflakeWriterActor.remote(table_name, connection_creds)  # type: ignore[attr-defined]
-                for _ in range(max_sessions)
-            ]
-        )
-        ray_df = pd.DataFrame(query_compiler=ray_qc)
-
-        original_column_labels = ray_df.columns.tolist()
-        original_column_index_names = ray_df.columns.names
-        data_column_snowflake_quoted_identifiers = (
-            generate_snowflake_quoted_identifiers_helper(
-                pandas_labels=original_column_labels, excluded=[]
-            )
-        )
-        ray_df.columns = [
-            extract_pandas_label_from_snowflake_quoted_identifier(identifier)
-            for identifier in data_column_snowflake_quoted_identifiers
-        ]
-        original_index_pandas_labels = ray_df.index.names
-        index_snowflake_quoted_identifiers = (
-            generate_snowflake_quoted_identifiers_helper(
-                pandas_labels=fill_none_in_index_labels(original_index_pandas_labels),
-                excluded=data_column_snowflake_quoted_identifiers,
-                wrap_double_underscore=True,
-            )
-        )
-        current_df_data_column_snowflake_quoted_identifiers = (
-            index_snowflake_quoted_identifiers
-            + data_column_snowflake_quoted_identifiers
-        )
-        index_names = [
-            extract_pandas_label_from_snowflake_quoted_identifier(identifier)
-            for identifier in index_snowflake_quoted_identifiers
-        ]
-        ray_df.reset_index(
-            inplace=True,
-            allow_duplicates=True,
-            names=index_names,
-        )
-        row_position_snowflake_quoted_identifier = (
-            generate_snowflake_quoted_identifiers_helper(
-                pandas_labels=[ROW_POSITION_COLUMN_LABEL],
-                excluded=current_df_data_column_snowflake_quoted_identifiers,
-                wrap_double_underscore=True,
-            )[0]
-        )
-        row_position_pandas_label = (
-            extract_pandas_label_from_snowflake_quoted_identifier(
-                row_position_snowflake_quoted_identifier
-            )
-        )
-
-        ray_df[row_position_pandas_label] = np.arange(len(ray_df))
-
-        current_df_data_column_snowflake_quoted_identifiers.append(
-            row_position_snowflake_quoted_identifier
-        )
-
-        with config_context(Backend="Ray"):
-            ray_ds = pd.io.to_ray(ray_df)
-        # Wait for all actors to finish writing
-        list(
-            pool.map_unordered(
-                lambda a, v: a.write.remote(v),
-                ray_ds.iter_batches(batch_size=None, batch_format="pandas"),
-            )
-        )
-
-        with config_context(Backend="Snowflake"):
-            snowpark_pandas_df = pd.read_snowflake(
-                table_name, index_col=index_names, enforce_ordering=True
-            )
-
-        snowpark_pandas_df.sort_values(by=row_position_pandas_label, inplace=True)
-        snowpark_pandas_df.drop(row_position_pandas_label, axis=1, inplace=True)
-        # Drop the permanent table and reference only the snapshot
-        pd.session.sql(f"DROP TABLE IF EXISTS {table_name}").collect()
-
-        return cls(
-            InternalFrame.create(
-                ordered_dataframe=snowpark_pandas_df._query_compiler._modin_frame.ordered_dataframe,
-                data_column_pandas_labels=original_column_labels,
-                data_column_pandas_index_names=original_column_index_names,
-                data_column_snowflake_quoted_identifiers=data_column_snowflake_quoted_identifiers,
-                index_column_pandas_labels=original_index_pandas_labels,
-                index_column_snowflake_quoted_identifiers=index_snowflake_quoted_identifiers,
-                data_column_types=None,
-                index_column_types=None,
-            )
-        )
 
     @classmethod
     def move_from(
@@ -1460,7 +1230,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
     ) -> Union[BaseQueryCompiler, Any]:  # pragma: no cover
         try:
             if source_qc.get_backend() == "Ray":
-                return cls._move_from_ray(source_qc, max_sessions=8)
+                return move_from_ray_helper(source_qc, max_sessions=8)
             return NotImplemented
         except Exception:
             return NotImplemented
